+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 0.0.5
+
+* Remove support for horizontal scroll direction
+* Change ListView to SliverAppBar
+* Add slivers parameters in CustomScrollView
+
 ## 0.0.4
 
-* Support for Web Application
+* Support for Web Applicationy
 
 ## 0.0.3